+++ nym-connect/src-tauri/src/operations/connection/connect.rs
@@ -18,3 +18,45 @@ pub async fn start_connecting(
         address: "Test".to_string(),
     })
 }
+
+#[tauri::command]
+pub async fn get_service_provider(
+    state: tauri::State<'_, Arc<RwLock<State>>>,
+) -> Result<String, BackendError> {
+    let guard = state.read().await;
+    guard
+        .get_service_provider()
+        .clone()
+        .ok_or(BackendError::NoServiceProviderSet)
+}
+
+#[tauri::command]
+pub async fn set_service_provider(
+    service_provider: String,
+    state: tauri::State<'_, Arc<RwLock<State>>>,
+) -> Result<(), BackendError> {
+    let mut guard = state.write().await;
+    guard.set_service_provider(service_provider);
+    Ok(())
+}
+
+#[tauri::command]
+pub async fn get_gateway(
+    state: tauri::State<'_, Arc<RwLock<State>>>,
+) -> Result<String, BackendError> {
+    let guard = state.read().await;
+    guard
+        .get_gateway()
+        .clone()
+        .ok_or(BackendError::NoGatewaySet)
+}
+
+#[tauri::command]
+pub async fn set_gateway(
+    gateway: String,
+    state: tauri::State<'_, Arc<RwLock<State>>>,
+) -> Result<(), BackendError> {
+    let mut guard = state.write().await;
+    guard.set_gateway(gateway);
+    Ok(())
+}