+++ nym-wallet/src/components/Rewards/CompoundModal.tsx
@@ -1,24 +1,39 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { Stack, Typography } from '@mui/material';
 import { IdentityKeyFormField } from '@nymproject/react/mixnodes/IdentityKeyFormField';
+import { simulateCompoundDelgatorReward, simulateVestingCompoundDelgatorReward } from 'src/requests';
+import { useGetFee } from 'src/hooks/useGetFee';
 import { SimpleModal } from '../Modals/SimpleModal';
+import { ModalFee } from '../Modals/ModalFee';
+import { FeeDetails } from '@nymproject/types';
 import { FeeWarning } from '../FeeWarning';
 
 export const CompoundModal: React.FC<{
   open: boolean;
   onClose?: () => void;
-  onOk?: (identityKey: string) => void;
+  onOk?: (identityKey: string, fee?: FeeDetails) => void;
   identityKey: string;
   amount: number;
-  fee: number;
+  minimum?: number;
   currency: string;
   message: string;
-}> = ({ open, onClose, onOk, identityKey, amount, fee, currency, message }) => {
-  const handleOk = () => {
+  usesVestingTokens: boolean;
+}> = ({ open, onClose, onOk, identityKey, amount, currency, message, usesVestingTokens }) => {
+  const { fee, isFeeLoading, feeError, getFee } = useGetFee();
+
+  const handleOk = async () => {
     if (onOk) {
-      onOk(identityKey);
+      onOk(identityKey, fee);
     }
   };
+
+  useEffect(() => {
+    if (usesVestingTokens) getFee(simulateVestingCompoundDelgatorReward, identityKey);
+    else {
+      getFee(simulateCompoundDelgatorReward, identityKey);
+    }
+  }, []);
+
   return (
     <SimpleModal
       open={open}
@@ -49,7 +64,8 @@ export const CompoundModal: React.FC<{
           {fee} {currency}
         </Typography>
       </Stack>
-      <FeeWarning amount={amount} fee={fee} />
+      {fee && <FeeWarning amount={amount} fee={fee} />}
+      <ModalFee fee={fee} isLoading={isFeeLoading} error={feeError} />
     </SimpleModal>
   );
 };