+++ validator-api/validator-api-requests/src/lib.rs
@@ -1,4 +1,6 @@
 // Copyright 2022 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
+#[cfg(feature = "coconut")]
+pub mod coconut;
 pub mod models;